// src/components/typography/enums.ts
export enum TextSize {
  DisplayXxl = "7.2",
  DisplayXl = "6",
  DisplayLg = "4.8",
  DisplayMd = "3.6",
  DisplaySm = "3",
  DisplayXs = "2.4",
  xl = "2",
  lg = "1.8",
  md = "1.6",
  sm = "1.4",
  xs = "1.2"
}

export enum TextWeight {
  regular = "400",
  medium = "500",
  semibold = "600",
  bold = "700"
}

export enum TextColor {
  base = "rgb(16,24,40)",
  white = "rgb(255,255,255)"
}

export enum TextAlign {
  center = "center",
  end = "end",
  justify = "justify",
  left = "left",
  right = "right",
  start = "start"
}